# packages/sdk/openapi.json
{"openapi":"3.1.0","info":{"title":"Vocal API","description":"Generic Speech AI Platform (STT + TTS)","version":"0.1.0"},"paths":{"/v1/audio/transcriptions":{"post":{"tags":["transcription"],"summary":"Transcribe audio","description":"Transcribe audio file to text using specified model","operationId":"create_transcription_v1_audio_transcriptions_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_transcription_v1_audio_transcriptions_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TranscriptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/audio/translations":{"post":{"tags":["transcription"],"summary":"Translate audio to English","description":"Translate audio to English text","operationId":"create_translation_v1_audio_translations_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_translation_v1_audio_translations_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TranscriptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/models":{"get":{"tags":["models"],"summary":"List models","description":"List all available models (Ollama-style)","operationId":"list_models_v1_models_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"task","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Task"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/models/{model_id}":{"get":{"tags":["models"],"summary":"Get model info","description":"Get detailed information about a specific model","operationId":"get_model_v1_models__model_id__get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["models"],"summary":"Delete model","description":"Remove a downloaded model","operationId":"delete_model_v1_models__model_id__delete","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/models/{model_id}/download":{"post":{"tags":["models"],"summary":"Download model","description":"Download a model for local use (Ollama-style pull)","operationId":"download_model_v1_models__model_id__download_post","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelDownloadProgress"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/models/{model_id}/download/status":{"get":{"tags":["models"],"summary":"Get download status","description":"Check model download progress","operationId":"get_download_status_v1_models__model_id__download_status_get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelDownloadProgress"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["health"],"summary":"Root","description":"API health check","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","description":"Detailed health check","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_create_transcription_v1_audio_transcriptions_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"Audio file to transcribe"},"model":{"type":"string","title":"Model","description":"Model ID","default":"Systran/faster-whisper-tiny"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language","description":"Language code"},"prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Prompt","description":"Style prompt"},"response_format":{"$ref":"#/components/schemas/TranscriptionFormat","description":"Output format","default":"json"},"temperature":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Temperature","default":0.0}},"type":"object","required":["file"],"title":"Body_create_transcription_v1_audio_transcriptions_post"},"Body_create_translation_v1_audio_translations_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"model":{"type":"string","title":"Model","default":"Systran/faster-whisper-tiny"}},"type":"object","required":["file"],"title":"Body_create_translation_v1_audio_translations_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ModelBackend":{"type":"string","enum":["faster_whisper","transformers","ctranslate2","nemo","onnx","custom"],"title":"ModelBackend","description":"Model inference backend"},"ModelDownloadProgress":{"properties":{"model_id":{"type":"string","title":"Model Id"},"status":{"type":"string","title":"Status"},"progress":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Progress","default":0.0},"downloaded_bytes":{"type":"integer","title":"Downloaded Bytes","default":0},"total_bytes":{"type":"integer","title":"Total Bytes","default":0},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"}},"type":"object","required":["model_id","status"],"title":"ModelDownloadProgress","description":"Model download progress"},"ModelInfo":{"properties":{"id":{"type":"string","title":"Id","description":"Unique model identifier"},"name":{"type":"string","title":"Name","description":"Human-readable model name"},"provider":{"$ref":"#/components/schemas/ModelProvider","description":"Model provider"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"size":{"type":"integer","title":"Size","description":"Model size in bytes","default":0},"size_readable":{"type":"string","title":"Size Readable","description":"Human-readable size","default":"Unknown"},"parameters":{"type":"string","title":"Parameters","description":"Number of parameters","default":"Unknown"},"languages":{"items":{"type":"string"},"type":"array","title":"Languages","description":"Supported languages"},"backend":{"$ref":"#/components/schemas/ModelBackend","description":"Inference backend"},"status":{"$ref":"#/components/schemas/ModelStatus","description":"Current model status"},"source_url":{"anyOf":[{"type":"string","maxLength":2083,"minLength":1,"format":"uri"},{"type":"null"}],"title":"Source Url"},"license":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"License"},"recommended_vram":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Recommended Vram"},"task":{"$ref":"#/components/schemas/ModelTask","description":"Task type"},"local_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Local Path"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","name","provider","backend","status","task"],"title":"ModelInfo","description":"Model information schema"},"ModelListResponse":{"properties":{"models":{"items":{"$ref":"#/components/schemas/ModelInfo"},"type":"array","title":"Models"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["models","total"],"title":"ModelListResponse","description":"List of available models"},"ModelProvider":{"type":"string","enum":["huggingface","local","custom"],"title":"ModelProvider","description":"Model provider/source"},"ModelStatus":{"type":"string","enum":["available","downloading","not_downloaded","error"],"title":"ModelStatus","description":"Model download/availability status"},"ModelTask":{"type":"string","enum":["stt","tts"],"title":"ModelTask","description":"Model task type"},"TranscriptionFormat":{"type":"string","enum":["json","text","srt","vtt","verbose_json"],"title":"TranscriptionFormat","description":"Output format for transcription"},"TranscriptionResponse":{"properties":{"text":{"type":"string","title":"Text","description":"Full transcribed text"},"language":{"type":"string","title":"Language","description":"Detected or specified language"},"duration":{"type":"number","title":"Duration","description":"Audio duration in seconds"},"segments":{"anyOf":[{"items":{"$ref":"#/components/schemas/TranscriptionSegment"},"type":"array"},{"type":"null"}],"title":"Segments"},"words":{"anyOf":[{"items":{"$ref":"#/components/schemas/TranscriptionWord"},"type":"array"},{"type":"null"}],"title":"Words"}},"type":"object","required":["text","language","duration"],"title":"TranscriptionResponse","description":"Response schema for transcription","example":{"duration":2.5,"language":"en","segments":[{"end":2.5,"id":0,"start":0.0,"text":"Hello, how are you today?"}],"text":"Hello, how are you today?"}},"TranscriptionSegment":{"properties":{"id":{"type":"integer","title":"Id"},"start":{"type":"number","title":"Start","description":"Start time in seconds"},"end":{"type":"number","title":"End","description":"End time in seconds"},"text":{"type":"string","title":"Text","description":"Transcribed text"},"tokens":{"anyOf":[{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Tokens"},"temperature":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Temperature"},"avg_logprob":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Avg Logprob"},"compression_ratio":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Compression Ratio"},"no_speech_prob":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"No Speech Prob"}},"type":"object","required":["id","start","end","text"],"title":"TranscriptionSegment","description":"A segment of transcribed text with timing"},"TranscriptionWord":{"properties":{"word":{"type":"string","title":"Word"},"start":{"type":"number","title":"Start"},"end":{"type":"number","title":"End"},"probability":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Probability"}},"type":"object","required":["word","start","end"],"title":"TranscriptionWord","description":"Word-level timestamp"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}